api/requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# 딥러닝
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
opencv-python>=4.8.0
ultralytics>=8.0.0

# 유틸리티
python-dotenv==1.0.0

# 테스트
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
